requirements.txt
asgiref==3.5.2
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.9.4
pytz==2022.5
sqlparse==0.4.3
